Issue #13345: Rename Example classes
