+++ src/xdocs-examples/resources/com/puppycrawl/tools/checkstyle/checks/design/mutableexception/Example2.java
@@ -1,20 +1,20 @@
 /*xml
 <module name="Checker">
   <module name="TreeWalker">
-    <module name="MutableException"/>
+    <module name="MutableException">
+      <property name="format" value="^.*Exception$"/>
+    </module>
   </module>
 </module>
-
-
 */
 
-package com.puppycrawl.tools.checkstyle.checks.design.mutableexception.Example1;
+package com.puppycrawl.tools.checkstyle.checks.design.mutableexception;
 
 // xdoc section -- start
-class FirstClass extends Exception {
-    private int code; // OK, class name doesn't match with default pattern
+class Example2 extends Exception {
+    private int code; // OK, class name doesn't match with given pattern
 
-    public FirstClass() {
+    public Example2() {
         code = 1;
     }
 }
@@ -28,8 +28,8 @@ class MyException extends Exception {
 }
 
 class MyThrowable extends Throwable {
-    final int code; // OK
-    String message; // violation
+    final int code; // OK, class name doesn't match with given pattern
+    String message; // OK, class name doesn't match with given pattern
 
     public MyThrowable(int code, String message) {
         this.code = code;
