+++ src/xdocs/checks/design/mutableexception.xml.template
@@ -48,13 +48,13 @@
         </p>
         <macro name="example">
           <param name="path"
-                 value="resources/com/puppycrawl/tools/checkstyle/checks/design/mutableexception/Example1/Example1.java"/>
+                 value="resources/com/puppycrawl/tools/checkstyle/checks/design/mutableexception/Example1.java"/>
           <param name="type" value="config"/>
         </macro>
         <p id="Example1-code">Example:</p>
         <macro name="example">
           <param name="path"
-                 value="resources/com/puppycrawl/tools/checkstyle/checks/design/mutableexception/Example1/Example1.java"/>
+                 value="resources/com/puppycrawl/tools/checkstyle/checks/design/mutableexception/Example1.java"/>
           <param name="type" value="code"/>
         </macro>
 
@@ -64,13 +64,13 @@
         </p>
         <macro name="example">
           <param name="path"
-                 value="resources/com/puppycrawl/tools/checkstyle/checks/design/mutableexception/Example2/Example2.java"/>
+                 value="resources/com/puppycrawl/tools/checkstyle/checks/design/mutableexception/Example2.java"/>
           <param name="type" value="config"/>
         </macro>
         <p id="Example2-code">Example:</p>
         <macro name="example">
           <param name="path"
-                 value="resources/com/puppycrawl/tools/checkstyle/checks/design/mutableexception/Example2/Example2.java"/>
+                 value="resources/com/puppycrawl/tools/checkstyle/checks/design/mutableexception/Example2.java"/>
           <param name="type" value="code"/>
         </macro>
 
@@ -80,13 +80,13 @@
         </p>
         <macro name="example">
           <param name="path"
-                 value="resources/com/puppycrawl/tools/checkstyle/checks/design/mutableexception/Example3/Example3.java"/>
+                 value="resources/com/puppycrawl/tools/checkstyle/checks/design/mutableexception/Example3.java"/>
           <param name="type" value="config"/>
         </macro>
         <p id="Example3-code">Example:</p>
         <macro name="example">
           <param name="path"
-                 value="resources/com/puppycrawl/tools/checkstyle/checks/design/mutableexception/Example3/Example3.java"/>
+                 value="resources/com/puppycrawl/tools/checkstyle/checks/design/mutableexception/Example3.java"/>
           <param name="type" value="code"/>
         </macro>
       </subsection>
